+++ relay-metrics/src/bucket.rs
@@ -930,12 +930,12 @@ mod tests {
         let s = "transactions/foo:42|c";
         let timestamp = UnixTimestamp::from_secs(4711);
         let metric = Bucket::parse(s.as_bytes(), timestamp).unwrap();
-        insta::assert_debug_snapshot!(metric, @r###"
+        insta::assert_debug_snapshot!(metric, @r#"
         Bucket {
             timestamp: UnixTimestamp(4711),
             width: 0,
             name: MetricName(
-                "c:transactions/foo@none",
+                "c:unsupported/foo@none",
             ),
             value: Counter(
                 42.0,
@@ -947,7 +947,7 @@ mod tests {
                 extracted_from_indexed: false,
             },
         }
-        "###);
+        "#);
     }
 
     #[test]
@@ -963,12 +963,12 @@ mod tests {
         let s = "transactions/foo:17.5|d";
         let timestamp = UnixTimestamp::from_secs(4711);
         let metric = Bucket::parse(s.as_bytes(), timestamp).unwrap();
-        insta::assert_debug_snapshot!(metric, @r###"
+        insta::assert_debug_snapshot!(metric, @r#"
         Bucket {
             timestamp: UnixTimestamp(4711),
             width: 0,
             name: MetricName(
-                "d:transactions/foo@none",
+                "d:unsupported/foo@none",
             ),
             value: Distribution(
                 [
@@ -982,7 +982,7 @@ mod tests {
                 extracted_from_indexed: false,
             },
         }
-        "###);
+        "#);
     }
 
     #[test]
@@ -1016,12 +1016,12 @@ mod tests {
         let s = "transactions/foo:4267882815|s";
         let timestamp = UnixTimestamp::from_secs(4711);
         let metric = Bucket::parse(s.as_bytes(), timestamp).unwrap();
-        insta::assert_debug_snapshot!(metric, @r###"
+        insta::assert_debug_snapshot!(metric, @r#"
         Bucket {
             timestamp: UnixTimestamp(4711),
             width: 0,
             name: MetricName(
-                "s:transactions/foo@none",
+                "s:unsupported/foo@none",
             ),
             value: Set(
                 {
@@ -1035,7 +1035,7 @@ mod tests {
                 extracted_from_indexed: false,
             },
         }
-        "###);
+        "#);
     }
 
     #[test]
@@ -1073,12 +1073,12 @@ mod tests {
         let s = "transactions/foo:42|g";
         let timestamp = UnixTimestamp::from_secs(4711);
         let metric = Bucket::parse(s.as_bytes(), timestamp).unwrap();
-        insta::assert_debug_snapshot!(metric, @r###"
+        insta::assert_debug_snapshot!(metric, @r#"
         Bucket {
             timestamp: UnixTimestamp(4711),
             width: 0,
             name: MetricName(
-                "g:transactions/foo@none",
+                "g:unsupported/foo@none",
             ),
             value: Gauge(
                 GaugeValue {
@@ -1096,7 +1096,7 @@ mod tests {
                 extracted_from_indexed: false,
             },
         }
-        "###);
+        "#);
     }
 
     #[test]
@@ -1104,12 +1104,12 @@ mod tests {
         let s = "transactions/foo:25:17:42:220:85|g";
         let timestamp = UnixTimestamp::from_secs(4711);
         let metric = Bucket::parse(s.as_bytes(), timestamp).unwrap();
-        insta::assert_debug_snapshot!(metric, @r###"
+        insta::assert_debug_snapshot!(metric, @r#"
         Bucket {
             timestamp: UnixTimestamp(4711),
             width: 0,
             name: MetricName(
-                "g:transactions/foo@none",
+                "g:unsupported/foo@none",
             ),
             value: Gauge(
                 GaugeValue {
@@ -1127,7 +1127,7 @@ mod tests {
                 extracted_from_indexed: false,
             },
         }
-        "###);
+        "#);
     }
 
     #[test]