+++ relay-metrics/src/cogs.rs
@@ -38,7 +38,6 @@ where
 fn to_app_feature(ns: MetricNamespace) -> AppFeature {
     match ns {
         MetricNamespace::Sessions => AppFeature::MetricsSessions,
-        MetricNamespace::Transactions => AppFeature::MetricsTransactions,
         MetricNamespace::Spans => AppFeature::MetricsSpans,
         MetricNamespace::Custom => AppFeature::MetricsCustom,
         MetricNamespace::Unsupported => AppFeature::MetricsUnsupported,