+++ relay-metrics/src/utils.rs
@@ -17,8 +17,6 @@ pub fn tags_cost(tags: &BTreeMap<String, String>) -> usize {
 pub struct ByNamespace<T> {
     /// Value for the [`MetricNamespace::Sessions`] namespace.
     pub sessions: T,
-    /// Value for the [`MetricNamespace::Transactions`] namespace.
-    pub transactions: T,
     /// Value for the [`MetricNamespace::Spans`] namespace.
     pub spans: T,
     /// Value for the [`MetricNamespace::Custom`] namespace.
@@ -32,7 +30,6 @@ impl<T> ByNamespace<T> {
     pub fn get(&self, namespace: MetricNamespace) -> &T {
         match namespace {
             MetricNamespace::Sessions => &self.sessions,
-            MetricNamespace::Transactions => &self.transactions,
             MetricNamespace::Spans => &self.spans,
             MetricNamespace::Custom => &self.custom,
             MetricNamespace::Unsupported => &self.unsupported,
@@ -43,7 +40,6 @@ impl<T> ByNamespace<T> {
     pub fn get_mut(&mut self, namespace: MetricNamespace) -> &mut T {
         match namespace {
             MetricNamespace::Sessions => &mut self.sessions,
-            MetricNamespace::Transactions => &mut self.transactions,
             MetricNamespace::Spans => &mut self.spans,
             MetricNamespace::Custom => &mut self.custom,
             MetricNamespace::Unsupported => &mut self.unsupported,
@@ -53,12 +49,11 @@ impl<T> ByNamespace<T> {
 
 impl<T> IntoIterator for ByNamespace<T> {
     type Item = (MetricNamespace, T);
-    type IntoIter = std::array::IntoIter<(MetricNamespace, T), 5>;
+    type IntoIter = std::array::IntoIter<(MetricNamespace, T), 4>;
 
     fn into_iter(self) -> Self::IntoIter {
         let Self {
             sessions,
-            transactions,
             spans,
             custom,
             unsupported,
@@ -66,7 +61,6 @@ impl<T> IntoIterator for ByNamespace<T> {
 
         [
             (MetricNamespace::Sessions, sessions),
-            (MetricNamespace::Transactions, transactions),
             (MetricNamespace::Spans, spans),
             (MetricNamespace::Custom, custom),
             (MetricNamespace::Unsupported, unsupported),
@@ -109,14 +103,12 @@ macro_rules! impl_op {
             fn $opfn(&mut self, rhs: Self) {
                 let Self {
                     sessions,
-                    transactions,
                     spans,
                     custom,
                     unsupported,
                 } = self;
 
                 $op::$opfn(sessions, rhs.sessions);
-                $op::$opfn(transactions, rhs.transactions);
                 $op::$opfn(spans, rhs.spans);
                 $op::$opfn(custom, rhs.custom);
                 $op::$opfn(unsupported, rhs.unsupported);