+++ relay-server/src/services/processor.rs
@@ -44,8 +44,7 @@ use crate::extractors::{PartialDsn, RequestMeta, RequestTrust};
 use crate::integrations::Integration;
 use crate::managed::{InvalidProcessingGroupType, ManagedEnvelope, TypedEnvelope};
 use crate::metrics::{MetricOutcomes, MetricsLimiter, MinimalTrackableBucket};
-use crate::metrics_extraction::transactions::ExtractedMetrics;
-use crate::metrics_extraction::transactions::types::ExtractMetricsError;
+use crate::metrics_extraction::ExtractedMetrics;
 use crate::processing::attachments::AttachmentProcessor;
 use crate::processing::check_ins::CheckInsProcessor;
 use crate::processing::client_reports::ClientReportsProcessor;
@@ -597,9 +596,6 @@ pub enum ProcessingError {
     #[error("event filtered with reason: {0:?}")]
     EventFiltered(FilterStatKey),
 
-    #[error("missing or invalid required event timestamp")]
-    InvalidTimestamp,
-
     #[error("could not serialize event payload")]
     SerializeFailed(#[source] serde_json::Error),
 
@@ -643,7 +639,6 @@ impl ProcessingError {
             Self::UnsupportedSecurityType => Some(Outcome::Filtered(FilterStatKey::InvalidCsp)),
             Self::InvalidNelReport(_) => Some(Outcome::Invalid(DiscardReason::InvalidJson)),
             Self::InvalidTransaction => Some(Outcome::Invalid(DiscardReason::InvalidTransaction)),
-            Self::InvalidTimestamp => Some(Outcome::Invalid(DiscardReason::Timestamp)),
             Self::DuplicateItem(_) => Some(Outcome::Invalid(DiscardReason::DuplicateItem)),
             Self::NoEventPayload => Some(Outcome::Invalid(DiscardReason::NoEventPayload)),
             Self::InvalidNintendoDyingMessage(_) => Some(Outcome::Invalid(DiscardReason::Payload)),
@@ -686,16 +681,6 @@ impl From<Unreal4Error> for ProcessingError {
         match err.kind() {
             Unreal4ErrorKind::TooLarge => Self::PayloadTooLarge(ItemType::UnrealReport.into()),
             _ => ProcessingError::InvalidUnrealReport(err),
-        }
-    }
-}
-
-impl From<ExtractMetricsError> for ProcessingError {
-    fn from(error: ExtractMetricsError) -> Self {
-        match error {
-            ExtractMetricsError::MissingTimestamp | ExtractMetricsError::InvalidTimestamp => {
-                Self::InvalidTimestamp
-            }
         }
     }
 }
@@ -784,18 +769,11 @@ impl ProcessingExtractedMetrics {
         // flag reset to `false`.
         let mut reset_extracted_from_indexed: SmallVec<[_; 2]> = smallvec![];
 
-        for (namespace, limit, indexed) in [
-            (
-                MetricNamespace::Transactions,
-                &enforcement.event,
-                &enforcement.event_indexed,
-            ),
-            (
-                MetricNamespace::Spans,
-                &enforcement.spans,
-                &enforcement.spans_indexed,
-            ),
-        ] {
+        for (namespace, limit, indexed) in [(
+            MetricNamespace::Spans,
+            &enforcement.spans,
+            &enforcement.spans_indexed,
+        )] {
             if limit.is_active() {
                 drop_namespaces.push(namespace);
             } else if indexed.is_active() && !enforced_consistently {
@@ -2334,7 +2312,7 @@ impl EnvelopeProcessorService {
             let over_accept_once = true;
             let mut rate_limits = RateLimits::new();
 
-            for category in [DataCategory::Transaction, DataCategory::Span] {
+            for category in [DataCategory::Span] {
                 let count = bucket_limiter.count(category);
 
                 let timer = Instant::now();
@@ -3318,7 +3296,7 @@ mod tests {
 
             let project_metrics = |scoping| ProjectBuckets {
                 buckets: vec![Bucket {
-                    name: "d:transactions/bar".into(),
+                    name: "d:spans/bar".into(),
                     value: BucketValue::Counter(FiniteF64::new(1.0).unwrap()),
                     timestamp: UnixTimestamp::now(),
                     tags: Default::default(),
@@ -3687,10 +3665,7 @@ mod tests {
         .await;
 
         let mut item = Item::new(ItemType::Statsd);
-        item.set_payload(
-            ContentType::Text,
-            "transactions/foo:3182887624:4267882815|s",
-        );
+        item.set_payload(ContentType::Text, "spans/foo:3182887624:4267882815|s");
         for (source, expected_received_at) in [
             (
                 BucketSource::External,